requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
requests
python-dotenv
volcengine-python-sdk[ark]
Authlib